original/requirements.txt
torch
torchvision
onnx
onnxruntime
Pillow
numpy
matplotlib
opencv-python
requests
